mirage: Remove default scenario
Supporting mirage in development builds is difficult without `ember-cli-mirage` since without proper tree shaking we would be shipping the additional mirage dependency bytes to production as well. Since most development happens in `start:live` mode and with help of the test suite, the use of this default scenario is somewhat questionable anyway.