CI: Remove broken CodeCov integration
This has been silently broken for a while and it's unclear at this point what exactly broke it. Since we rarely look at it and the `grcov` call takes the majority of our CI time, we will remove this for now in the hope that a better alternative will emerge eventually.