Fix `remove_nonexistent_team()` test
This test was supposed to check what happens when a user removes a team owner that no longer exists on GitHub. But the test never added the team as a crate owner in the first place, which resulted in it not actually testing the relevant removal process. This commit fixes the issue by adding the corresponding row in the `crate_owners` database table.