Add a username field to the users table (deploy 3?)
Right now, this will always have the same value as gh_login on the users
table and login on the linked accounts table. All of these values will
get updated when we get a new gh_login value.

Eventually, we're going to have to decouple the concept of crates.io
"username" from the logins of the various account(s), which you may or
may not want to update when you update your GitHub or other login, and
which may or may not conflict with other users' crates.io usernames.

But for now, set up for that future and defer the harder decisions until
later by making this field always get the same value as gh_login.