Add a username field to the users table (deploy 3?)
Right now, this will always have the same value as gh_login on the users
table and login on the linked accounts table. All of these values will
get updated when we get a new gh_login value.

Eventually, we're going to have to decouple the concept of crates.io
"username" from the logins of the various account(s), which you may or
may not want to update when you update your GitHub or other login, and
which may or may not conflict with other users' crates.io usernames.

But for now, set up for that future and defer the harder decisions until
later by making this field always get the same value as gh_login.

This adds a `username` field to the JSON views returned from the API but
does not use that field anywhere yet.

Question: should team owner JSON also have a field named `username` as
done in this commit? it's a little weird for a team to have a username
because it's not a user. but it's consistent. something more generic
like `name` might also be confusing. something more specific like
`crates_io_identifier` is more verbose but less confusable. shrug