controllers/krate/search: Build query filter for seek-based pagination
...based on direction
Most of this should be intuitive, as it simply flips the comparison
operator. To make it easier to identify the difference, I chose to
implement it in a single function and place it next to the forward
ordering conditions. The only part that requires a bit of thought is the
case involving nulls last.