controllers/helpers/pagination: Maintain consistent ordering from page to page
For instance, given rows with the following:

| a | b | c | ... | 1 | 2 | 3 |

When performing forward seek pagination with page size as 3 to the last
page, let's say we have the last page as 1, 2, 3. When performing
backward seek pagination, we reverse the order of the query so that we
can retrieve the data from the tail. This results in 3, 2, 1. However,
we would like to maintain consistent ordering as we view it with forward
pagination, namely 1, 2, 3. Therefore, we need to reverse the results
again to maintain the original ordering.