Replace `ember-router-scroll` with `memory-scroll` (#10962)
The former hasn't been maintained for a couple of years, while the latter appears to be properly maintained. The behavior is ever so slightly different, but for the most part they both work roughly the same. The main purpose is to scroll the website back to the top whenever the user navigates to a new page.