insta: Update snapshots
Apparently the `zlib-rs` deflate implementation produces slightly different output to `miniz_oxide`, which results in the tarball checksums not matching up and the file sizes being off in some cases. AFAICT this is only relevant for the tarballs that are produced as fixtures by our own test suite and this doesn't actually influence the production code at all since the tarball reading code isn't affected.