tests: Replace `assert_eq!` with `assert_snapshot!` for status code checks (#11346)
This commit updates test assertions to use `insta::assert_snapshot!` for HTTP status code checks across various test files. This makes it easier to update the assertions in case status codes for certain errors are changed/fixed.

Unfortunately, this changes the numbering of the file-based snapshots, which causes the diff to look much larger than it actually is... 🙈