msw: Implement relevant Trusted Publishing endpoints
This commit implements the three `/api/v1/trusted_publishing/github_configs` endpoints (list, create, delete). It adds a dedicated `trustpubGithubConfig` model and corresponding serializer and then uses them to implement the endpoints. As usual with msw, this is a relatively simple implementation, but good enough for our frontend test suite.