trustpub: Prefill repo owner and name for crates with GitHub repo URLs
When the user visits the form to create a new Trusted Publishing configuration, we look at the `repository` field of the crate, and if it is a GitHub URL (e.g. https://github.com/rust-lang/crates.io) we extract the repository owner and name from the URL and prefill the input fields accordingly. This should make the configuration creation a slightly more enjoyable experience.