Add an API route for admins to assess crates
Connects to #10387.

This API route, under `/private/` rather than `v1`, will make it a lot
easier to assess whether crates contain useful functionality or are
squatting.

This is only the backend; some of the functionality described in the
issue will be frontend-only, but I wanted to get the API in first.

The API route will only return data if the currently authenticated user
is an admin. This is important because the crate owner's verified email
address is part of the returned data so that the admin can contact the
owner if necessary.

Another reason this route is limited to admins is that some of the
queries may be slow.