Fix unintentional `crate.newest_version` reset (#11795)
The `GET /summary` call correctly initialized the `newest_version` fields of the crates, but when navigating to the crate page, the `GET /crate/foo` call by default uses an `include` query parameter that excludes the `newest_version` field from being loaded on the backend, which causes it to be returned with a value of `0.0.0` (for legacy reasons). This then reset the field on the model, causing it to be displayed incorrectly when navigating back to the front page.

This commit fixes the issue by adjusting the crate serializer to ignore `0.0.0` values for `max_version` and `newest_version`, so that they won't override previously loaded values.