msw: Avoid unnecessary setup and teardown (#11884)
Instead of creating, starting and tearing down the worker for each test module, we can create it once, start it once for the full test suite and then tear it down at the end of the test suite. This also ensures that we don't accidentally perform network calls in tests that didn't explicitly set up MSW.

Additionally, this might fix the flakiness we've been seeing with the test suite when run on the CI runners, and is generally recommended by the MSW devs.