Always print download numbers when deleting a crate
So that if we're deleting it for security reasons but forgot to grab the
downloads from the page, the person deleting it at least has the
download numbers handy in their console.

Rather than making _display_ conditional on how high the download
numbers are, make the _red and bold_ styling conditional on how high the
download numbers are.