Fix dropdown indicator color in dark mode
Unfortunately using `currentColor` does not work in SVGs when they are used as a `background-image`. Using a CSS filter also didn't work since it targeted the whole control. Next I tried to use `light-dark`, but apparently that is not valid for `background-image` either, so I ended up with this slightly more complicated pattern. There might be ways to improve this, but this is good enough for now unless anyone wants to spend time optimizing this further :)