prettier: Disable for markdown files (#12333)
This is causing more pain than it is worth, especially since the prettier check currently only runs in the frontend test suite, so if a backend pull request introduces a new markdown file the CI will fail on some future run instead of on the PR. Sure, we can fix that, but is it worth the complexity? IMHO no.