Document that the backend assumes `cfg(unix)`
Currently the signal handling logic for triggering a graceful shutdown
fails to build on Windows. It would be easy to add Windows support here,
but given this has been broken for 4.5 years and nobody has complained
it seems there isn't much demand for this.

As far as I'm aware, frontend development on Windows should still be
supported, so I left the reference to Windows in that section.