Add monthly and weekly downloads to recent downloads calculation
We calculate the recent downloads using a materialized view that
we refresh every day. But it's only for last 90 days.

Most of the other ecosystems like NPM, PyPi, Packagist, Hex
provide daily, weekly and monthly downloads. So, I added
a weekly and monthly column in the recent downloads view
and surfaced the data in the `/api/v1/crates/{name}` endpoint.