ChaosProxy: accept then immediately drop connection
This slightly changes the low-level TCP behavior. The previous
implementation dropped the listener which results in an immediate TCP
RST in response to the client's SYN packet.  In the new implementation
the SYN handshake completes before a RST is sent.

This prepares for the next commit, which will change the TCP connection
to a Unix socket. Unfortunately, dropping the listener, manually
deleting the socket file, and then recreating the socket does not result
in a healthy connection. This commit splits off the slight semantic
change while still passing the tests over TCP sockets.

I investigated taking this further such that the connection is accepted
and stalled indefinitely (by pushing the stream to a Vec that is owned
for the duration of the test). This resulted in the chaos proxy tests
hanging until the timeout was hit. It also changed the status code
observed in the test to a 408 Request Timeout. (Most of these tests
currently return 503 Service Unavailable but
`fallback_to_replica_returns_user_info` is expected to return 200 OK.)
Longer term, maybe it would make sense to make the behvior more
consistent and to add test coverage for both failure modes.