ChaosProxy: Proxy over Unix socket instead of TCP
Each proxy now creates a temporary directory containing a Unix socket.
Connections made to the Unix socket are forward to the database backend
over whichever method is configured via `TEST_DATABASE_URL`.

The reason for this change is that otherwise if `TEST_DATABASE_URL` was
pointed to a Unix socket then while running tests another user on the
localhost might be able to connect to the test database without
requiring credentials. It is unlikely that this is a relevant threat
model for most developers of crates.io, however it still seems best to
not risk ever creating a Postgres TCP -> Unix socket proxy.

The downside of this change is that the test configuration no longer
matches the TCP environment used in production. An alternative would be
to fail the test if a bad configuration was requested. Another
alternative is to duplicate the proxy logic to support both socket
types. This option was explored, however the resulting code duplication
did not seem to be worth the effort.