typosquat: Fail gracefully if crate does not exist
Up until now the `check_typosquat` background job would fail when run on a crate name that had already been deleted again. This failure meant that the job would enter a retry cycle, that it couldn't escape unless someone else published a crate with the same name after the cooldown period.

This commit adjusts the background job to deal with this situation gracefully by finishing the job successfully if the crate no longer exists.