Remove route name assertions (#12577)
The internal state of the Ember.js router doesn't matter much to us. We only really care about the URL and what is displayed on the page.

This commit removes the corresponding assertions.

This also makes the tests more portable to other frameworks, since they now don't rely on Ember.js-specific things anymore.