Integrate `eslint-plugin-prefer-let`
Add `eslint-plugin-prefer-let` to both Ember and Svelte ESLint configs to enforce using `let` instead of `const` for non-top-level declarations.

`const` in JS instead the same as `let` without `mut` in Rust. `const` just means that the variable binding is immutable, but the object it refers to can still be mutated. Since this makes the distinction between `let` and `const` in JS essentially meaningless from a technical point of view, we may as well use these keywords in a more useful way: `let` for variables, and `const` for module-level constants.