Unconditionally check docs.rs status
Even if the crate has a documentation link specified, we want to show a
link to the source view of docs.rs. But we still want to check that
docs.rs has had a chance to build the crate's docs before we add the
link to browse the source, so we do need to make this request.